fix: move client component to its own file
